requirements.txt
# Requirements.txt - Dependências do projeto Valorize Web
# Para instalar: pip install -r requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.7
